move apply_body_map and merge_body_maps to different file.
